Move maven and terminal functions to util file
