+++ src/terminal/mavenTerminal.ts
@@ -3,23 +3,15 @@
 // Licensed under the MIT license.
 
 import { Disposable, Terminal, env, window } from "vscode";
-import { quarkusOutputChannel } from "./quarkusOutputChannel";
 import { Settings } from "./settings";
-import { executeCommand } from "./cpUtils";
+
+import { WindowsShellType, currentWindowsShell, getCommand, getCDCommand, setupEnvForWSL } from '../utils/shellUtils';
 
 export interface ITerminalOptions {
     addNewLine?: boolean;
     name: string;
     cwd?: string;
     env?: { [key: string]: string };
-}
-
-enum WindowsShellType {
-    CMD = "Command Prompt",
-    POWER_SHELL = "PowerShell",
-    GIT_BASH = "Git Bash",
-    WSL = "WSL Bash",
-    OTHERS = "Others"
 }
 
 class MavenTerminal implements Disposable {
@@ -60,19 +52,6 @@ class MavenTerminal implements Disposable {
         }
     }
 
-    public async formattedPathForTerminal(filepath: string): Promise<string> {
-        if (process.platform === "win32") {
-            switch (currentWindowsShell()) {
-                case WindowsShellType.WSL:
-                    return await toWslPath(filepath);
-                default:
-                    return filepath;
-            }
-        } else {
-            return filepath;
-        }
-    }
-
     public dispose(id?: string): void {
         if (id) {
             this.terminals[id].dispose();
@@ -82,103 +61,4 @@ class MavenTerminal implements Disposable {
     }
 }
 
-function getCommand(cmd: string): string {
-    if (process.platform === "win32") {
-        switch (currentWindowsShell()) {
-            case WindowsShellType.POWER_SHELL:
-                return `cmd /c ${cmd}`; // PowerShell
-            default:
-                return cmd; // others, try using common one.
-        }
-    } else {
-        return cmd;
-    }
-}
-
-async function getCDCommand(cwd: string): Promise<string> {
-    if (process.platform === "win32") {
-        switch (currentWindowsShell()) {
-            case WindowsShellType.GIT_BASH:
-                return `cd "${cwd.replace(/\\+$/, "")}"`; // Git Bash: remove trailing '\'
-            case WindowsShellType.POWER_SHELL:
-                // Escape '[' and ']' in PowerShell
-                // See: https://github.com/microsoft/vscode-maven/issues/324
-                const escaped: string = cwd.replace(/([\[\]])/g, "``$1");
-                return `cd "${escaped}"`; // PowerShell
-            case WindowsShellType.CMD:
-                return `cd /d "${cwd}"`; // CMD
-            case WindowsShellType.WSL:
-                return `cd "${await toWslPath(cwd)}"`; // WSL
-            default:
-                return `cd "${cwd}"`; // Unknown, try using common one.
-        }
-    } else {
-        return `cd "${cwd}"`;
-    }
-}
-
-function currentWindowsShell(): WindowsShellType {
-    const currentWindowsShellPath: string = env.shell;
-
-    if (currentWindowsShellPath.endsWith("cmd.exe")) {
-        return WindowsShellType.CMD;
-    } else if (currentWindowsShellPath.endsWith("powershell.exe")) {
-        return WindowsShellType.POWER_SHELL;
-    } else if (currentWindowsShellPath.endsWith("bash.exe") || currentWindowsShellPath.endsWith("wsl.exe")) {
-        if (currentWindowsShellPath.includes("Git")) {
-            return WindowsShellType.GIT_BASH;
-        }
-        return WindowsShellType.WSL;
-    } else {
-        return WindowsShellType.OTHERS;
-    }
-}
-
-function toDefaultWslPath(p: string): string {
-    const arr: string[] = p.split(":\\");
-    if (arr.length === 2) {
-        const drive: string = arr[0].toLowerCase();
-        const dir: string = arr[1].replace(/\\/g, "/");
-        return `/mnt/${drive}/${dir}`;
-    } else {
-        return p.replace(/\\/g, "/");
-    }
-}
-
-export async function toWslPath(path: string): Promise<string> {
-    try {
-        return (await executeCommand("wsl", ["wslpath", "-u", `"${path.replace(/\\/g, "/")}"`])).trim();
-    } catch (error) {
-			quarkusOutputChannel.appendLine(error, "WSL");
-        return toDefaultWslPath(path);
-    }
-}
-
-export async function toWinPath(path: string): Promise<string> {
-    return (await executeCommand("wsl", ["wslpath", "-w", `"${path}"`])).trim();
-}
-
-export function getMavenWrapperFilename(): string {
-    if (process.platform === "win32") {
-        return 'mvnw.cmd';
-    }
-
-    return 'mvnw';
-}
-
-export function getMavenWrapperExecuteable(): string {
-	if (process.platform === "win32") {
-		return '.\\mvnw.cmd';
-	}
-	return './mvnw';
-}
-
 export const mavenTerminal: MavenTerminal = new MavenTerminal();
-
-function setupEnvForWSL(terminal: Terminal, env: { [envKey: string]: string }): void {
-    if (terminal) {
-        Object.keys(env).forEach(key => {
-            terminal.sendText(`export ${key}="${env[key]}"`, true);
-        });
-    }
-}
