+++ src/utils/cpUtils.ts
@@ -3,7 +3,7 @@
 
 import * as cp from "child_process";
 import * as vscode from "vscode";
-import { quarkusOutputChannel } from "./quarkusOutputChannel";
+import { quarkusOutputChannel } from "../terminal/quarkusOutputChannel";
 
 export async function executeCommand(command: string, args: string[], options: cp.SpawnOptions = { shell: true }): Promise<string> {
     return new Promise((resolve: (res: string) => void, reject: (e: Error) => void): void => {
