+++ vscode-debug-files/tasks.json
@@ -2,32 +2,5 @@
 // for the documentation about the tasks.json format
 {
 	"version": "2.0.0",
-	"tasks": [
-		{
-			"label": "quarkus:dev",
-			"type": "shell",
-			"command": "./mvnw quarkus:dev",
-			"windows": {
-				"command": ".\\mvn.cmd quarkus:dev"
-			},
-			"isBackground": true,
-			"problemMatcher": [
-				{
-					"pattern": [
-						{
-							"regexp": "\\b\\B",
-							"file": 1,
-							"location": 2,
-							"message": 3
-						}
-					],
-					"background": {
-						"activeOnStart": true,
-						"beginsPattern": "^.*Scanning for projects...*",
-						"endsPattern": "^.*Quarkus .* started in .*\\. Listening on:*"
-					}
-				}
-			]
-		}
-	]
+	"tasks": []
 }
